fix(parts): unify PO receiving and materialize request parts
