+++ features/parts/components/ScanToReceivePanel.tsx
@@ -1,0 +1,445 @@
+// app/parts/receive/page.tsx (FULL FILE REPLACEMENT)
+
+"use client";
+
+import { useEffect, useMemo, useRef, useState } from "react";
+import { createBrowserSupabase } from "@/features/shared/lib/supabase/client";
+import type { Database } from "@shared/types/types/supabase";
+import { resolveScannedCode } from "@/features/parts/server/scanActions";
+import Link from "next/link";
+import {
+  buildPartTrustMeta,
+  trustBadgeTone,
+  trustLevelLabel,
+  type PartTrustMeta,
+} from "@/features/parts/lib/trust-signals";
+import { desktopPrimitives as ui } from "@/features/shared/components/ui/desktopPrimitives";
+
+type QuaggaResult = { codeResult?: { code?: string | null } | null };
+
+type QuaggaConfig = {
+  inputStream: {
+    name: string;
+    type: "LiveStream";
+    target: HTMLElement;
+    constraints?: { facingMode?: string };
+  };
+  decoder: { readers: string[] };
+  locate?: boolean;
+};
+
+type QuaggaDetectedHandler = (res: QuaggaResult) => void;
+
+type QuaggaLike = {
+  init: (cfg: QuaggaConfig, cb: (err?: Error) => void) => void;
+  start: () => void;
+  stop: () => void;
+  onDetected: (handler: QuaggaDetectedHandler) => void;
+  offDetected?: (handler: QuaggaDetectedHandler) => void;
+};
+
+let Quagga: QuaggaLike | null = null;
+
+if (typeof window !== "undefined") {
+  void import("@ericblade/quagga2").then((m) => {
+    Quagga = (m.default as unknown) as QuaggaLike;
+  });
+}
+
+type DB = Database;
+type PurchaseOrder = DB["public"]["Tables"]["purchase_orders"]["Row"];
+type StockLoc = DB["public"]["Tables"]["stock_locations"]["Row"];
+type PartTrustFields = Pick<
+  DB["public"]["Tables"]["parts"]["Row"],
+  "id" | "sku" | "part_number" | "normalized_part_key" | "source_intake_id"
+>;
+
+type ReceiveResult =
+  | {
+      ok: true;
+      mode: "po" | "manual";
+      result?: unknown;
+    }
+  | { error: string };
+
+export function ScanToReceivePanel(): JSX.Element {
+  const supabase = useMemo(() => createBrowserSupabase(), []);
+  const [shopId, setShopId] = useState<string>("");
+
+  const [pos, setPOs] = useState<PurchaseOrder[]>([]);
+  const [selectedPo, setSelectedPo] = useState<string>("");
+  const [selectedLoc, setSelectedLoc] = useState<string>("");
+  const [locs, setLocs] = useState<StockLoc[]>([]);
+  const videoRef = useRef<HTMLDivElement | null>(null);
+  const [scanning, setScanning] = useState<boolean>(false);
+  const [qty, setQty] = useState<number>(1);
+
+  const [lastResult, setLastResult] = useState<ReceiveResult | null>(null);
+  const [lastTrust, setLastTrust] = useState<PartTrustMeta | null>(null);
+  const [busy, setBusy] = useState(false);
+
+  const onDetectedRef = useRef<QuaggaDetectedHandler | null>(null);
+  const lastScanRef = useRef("");
+  const receiveBusyRef = useRef(false);
+  const scanOperationRef = useRef<{ code: string; id: string } | null>(null);
+
+  useEffect(() => {
+    (async () => {
+      const { data: userRes } = await supabase.auth.getUser();
+      const uid = userRes.user?.id;
+      if (!uid) return;
+
+      // ✅ profiles.user_id == auth.uid()
+      const { data: prof, error: profErr } = await supabase
+        .from("profiles")
+        .select("shop_id")
+        .eq("user_id", uid)
+        .maybeSingle();
+
+      if (profErr) return;
+
+      const sid = prof?.shop_id ?? "";
+      setShopId(sid);
+      if (!sid) return;
+
+      const [poRes, locRes] = await Promise.all([
+        supabase
+          .from("purchase_orders")
+          .select("*")
+          .eq("shop_id", sid)
+          .order("created_at", { ascending: false })
+          .limit(50),
+        supabase
+          .from("stock_locations")
+          .select("*")
+          .eq("shop_id", sid)
+          .order("code"),
+      ]);
+
+      setPOs((poRes.data ?? []) as PurchaseOrder[]);
+      const locRows = (locRes.data ?? []) as StockLoc[];
+      setLocs(locRows);
+
+      const main = locRows.find((l) => (l.code ?? "").toUpperCase() === "MAIN");
+      if (main) setSelectedLoc(main.id);
+    })();
+  }, [supabase]);
+
+  const cleanupScannerHandlers = () => {
+    try {
+      if (Quagga && onDetectedRef.current && typeof Quagga.offDetected === "function") {
+        Quagga.offDetected(onDetectedRef.current);
+      }
+    } catch {
+      // ignore
+    }
+    onDetectedRef.current = null;
+  };
+
+  const startScan = async () => {
+    if (!Quagga || scanning || !videoRef.current) return;
+
+    setScanning(true);
+
+    Quagga.init(
+      {
+        inputStream: {
+          name: "Live",
+          type: "LiveStream",
+          target: videoRef.current,
+          constraints: { facingMode: "environment" },
+        },
+        decoder: {
+          readers: [
+            "upc_reader",
+            "upc_e_reader",
+            "ean_reader",
+            "ean_8_reader",
+            "code_128_reader",
+          ],
+        },
+        locate: true,
+      },
+      (err?: Error) => {
+        if (err) {
+          console.error(err);
+          setScanning(false);
+          return;
+        }
+        Quagga?.start();
+      },
+    );
+
+    cleanupScannerHandlers();
+
+    const handler: QuaggaDetectedHandler = async (res) => {
+      const code = res.codeResult?.code ?? "";
+      if (
+        !code ||
+        code === lastScanRef.current ||
+        receiveBusyRef.current
+      ) return;
+
+      lastScanRef.current = code;
+      receiveBusyRef.current = true;
+      setLastResult(null);
+
+      if (!selectedLoc) {
+        setLastResult({ error: "Select a location first." });
+        receiveBusyRef.current = false;
+        window.setTimeout(() => {
+          lastScanRef.current = "";
+        }, 800);
+        return;
+      }
+
+      const supplierId = pos.find((p) => p.id === selectedPo)?.supplier_id ?? null;
+
+      let part_id: string | null = null;
+      try {
+        ({ part_id } = await resolveScannedCode({
+          code,
+          supplier_id: supplierId,
+        }));
+      } catch (error) {
+        receiveBusyRef.current = false;
+        setLastResult({
+          error: error instanceof Error ? error.message : "Part lookup failed",
+        });
+        window.setTimeout(() => {
+          lastScanRef.current = "";
+        }, 900);
+        return;
+      }
+
+      if (!part_id) {
+        setLastResult({
+          error: `No part found for "${code}". Map it in Parts → Inventory → Edit → Barcodes.`,
+        });
+        receiveBusyRef.current = false;
+        scanOperationRef.current = null;
+        window.setTimeout(() => {
+          lastScanRef.current = "";
+        }, 1200);
+        return;
+      }
+      const { data: partRow } = await supabase
+        .from("parts")
+        .select("id, sku, part_number, normalized_part_key, source_intake_id")
+        .eq("id", part_id)
+        .maybeSingle();
+      const trustPart = (partRow ?? null) as PartTrustFields | null;
+      setLastTrust(
+        buildPartTrustMeta({
+          sku: trustPart?.sku ?? null,
+          partNumber: trustPart?.part_number ?? null,
+          normalizedPartKey: trustPart?.normalized_part_key ?? null,
+          sourceIntakeId: trustPart?.source_intake_id ?? null,
+        }),
+      );
+
+      const operationId =
+        scanOperationRef.current?.code === code
+          ? scanOperationRef.current.id
+          : crypto.randomUUID();
+      scanOperationRef.current = { code, id: operationId };
+      setBusy(true);
+      try {
+        const resp = await fetch("/api/receive-scan", {
+          method: "POST",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({
+            part_id,
+            location_id: selectedLoc,
+            qty,
+            po_id: selectedPo || null,
+            operation_id: operationId,
+          }),
+        });
+
+        const json = (await resp.json().catch(() => ({}))) as
+          | { ok?: boolean; mode?: "po" | "manual"; result?: unknown; error?: string }
+          | Record<string, unknown>;
+
+        if (!resp.ok) {
+          const errMsg =
+            typeof (json as { error?: unknown }).error === "string"
+              ? (json as { error: string }).error
+              : "Receive failed";
+          setLastResult({ error: errMsg });
+          return;
+        }
+
+        setLastResult({
+          ok: true,
+          mode: (json as { mode?: "po" | "manual" }).mode ?? (selectedPo ? "po" : "manual"),
+          result: (json as { result?: unknown }).result,
+        });
+
+        scanOperationRef.current = null;
+        window.dispatchEvent(new CustomEvent("parts:received"));
+      } catch (error) {
+        setLastResult({
+          error: error instanceof Error ? error.message : "Receive failed",
+        });
+      } finally {
+        receiveBusyRef.current = false;
+        setBusy(false);
+        window.setTimeout(() => {
+          lastScanRef.current = "";
+        }, 900);
+      }
+    };
+
+    onDetectedRef.current = handler;
+    Quagga.onDetected(handler);
+  };
+
+  const stopScan = () => {
+    try {
+      cleanupScannerHandlers();
+      Quagga?.stop();
+    } catch {
+      /* ignore */
+    }
+    setScanning(false);
+  };
+
+  useEffect(() => {
+    return () => stopScan();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  const locLabel = locs.find((l) => l.id === selectedLoc)?.code ?? "LOC";
+
+  return (
+    <section className="space-y-4" aria-labelledby="scan-to-receive-title">
+      <div>
+        <h2 id="scan-to-receive-title" className="text-lg font-semibold text-[color:var(--theme-text-primary)]">Scan to receive</h2>
+        <p className="mt-1 text-sm text-[color:var(--theme-text-secondary)]">Fast intake by barcode into a PO or stock location.</p>
+      </div>
+      <div className="space-y-4">
+
+        <div className="flex flex-wrap items-center gap-2">
+          <Link
+            href="/assistant?pageType=scan_receive&pageTitle=Scan%20to%20Receive"
+            className="rounded border border-sky-500/35 bg-sky-950/20 px-3 py-2 text-sm text-sky-200 hover:bg-sky-900/25"
+          >
+            Ask Assistant
+          </Link>
+        </div>
+      </div>
+
+      <div className="text-xs text-[color:var(--theme-text-muted)]">
+        Shop: <span className="text-[color:var(--theme-text-secondary)]">{shopId ? shopId.slice(0, 8) : "—"}</span>
+      </div>
+
+      <div className="desktop-toolbar-row grid gap-3 p-3 sm:grid-cols-3">
+        <div className="sm:col-span-1">
+          <div className="text-xs text-[color:var(--theme-text-secondary)] mb-1">Purchase Order (optional)</div>
+          <select
+            className={ui.input}
+            value={selectedPo}
+            onChange={(e) => setSelectedPo(e.target.value)}
+          >
+            <option value="">— No PO —</option>
+            {pos.map((po) => (
+              <option key={po.id} value={po.id}>
+                {po.id.slice(0, 8)} • {po.status ?? "draft"}
+              </option>
+            ))}
+          </select>
+        </div>
+
+        <div>
+          <div className="text-xs text-[color:var(--theme-text-secondary)] mb-1">Location</div>
+          <select
+            className={ui.input}
+            value={selectedLoc}
+            onChange={(e) => setSelectedLoc(e.target.value)}
+          >
+            {locs.map((l) => (
+              <option key={l.id} value={l.id}>
+                {l.code ?? "LOC"} — {l.name ?? ""}
+              </option>
+            ))}
+          </select>
+        </div>
+
+        <div>
+          <div className="text-xs text-[color:var(--theme-text-secondary)] mb-1">Quantity</div>
+          <input
+            type="number"
+            min={0.01}
+            step="0.01"
+            className={ui.input}
+            value={qty}
+            onChange={(e) => setQty(Math.max(0.01, Number(e.target.value || 1)))}
+          />
+        </div>
+      </div>
+
+      {/* Result panel */}
+      <div className="desktop-panel-soft p-3">
+        <div className="text-xs uppercase tracking-[0.18em] text-[color:var(--theme-text-muted)] mb-2">
+          Last receive
+        </div>
+
+        {!lastResult ? (
+          <div className="text-sm text-[color:var(--theme-text-secondary)]">
+            Scan a barcode to receive into <span className="text-[color:var(--theme-text-primary)]">{locLabel}</span>.
+          </div>
+        ) : "error" in lastResult ? (
+          <div className="text-sm text-red-300">{lastResult.error}</div>
+        ) : (
+          <div className="space-y-1">
+            <div className="text-sm text-emerald-300">
+              Received ×{qty} to {locLabel} ({lastResult.mode === "po" ? "PO receive + allocate" : "manual receive"})
+            </div>
+
+            {lastResult.mode === "po" ? (
+              <pre className="mt-2 max-h-64 overflow-auto rounded border border-[color:var(--theme-border-soft)] bg-[color:var(--theme-surface-overlay)] p-2 text-xs text-[color:var(--theme-text-primary)]">
+                {JSON.stringify(lastResult.result ?? {}, null, 2)}
+              </pre>
+            ) : null}
+            {lastTrust ? (
+              <div className="mt-1 text-xs">
+                <span className={`inline-flex rounded-full border px-2 py-1 ${trustBadgeTone(lastTrust.level)}`}>Trust: {trustLevelLabel(lastTrust.level)}</span>
+                {lastTrust.reasons.length > 0 ? <span className="ml-2 text-sky-200">{lastTrust.reasons.slice(0, 2).join(" · ")}</span> : null}
+              </div>
+            ) : null}
+          </div>
+        )}
+      </div>
+
+      <div className="desktop-panel-soft p-3">
+        <div className="mb-2 flex items-center gap-2">
+          {!scanning ? (
+            <button
+              onClick={startScan}
+              disabled={busy}
+              className="rounded border border-sky-500/35 px-3 py-1.5 text-sm text-sky-200 hover:bg-sky-900/20 disabled:opacity-60"
+            >
+              Start Scanner
+            </button>
+          ) : (
+            <button
+              onClick={stopScan}
+              className="rounded border border-red-500 px-3 py-1.5 text-sm text-red-300 hover:bg-red-900/20"
+            >
+              Stop Scanner
+            </button>
+          )}
+          <span className="text-xs text-[color:var(--theme-text-secondary)]">
+            Use mobile camera to scan UPC/EAN/Code128.
+          </span>
+        </div>
+
+        <div
+          ref={videoRef}
+          className="aspect-video w-full overflow-hidden rounded border border-[color:var(--theme-border-soft)] bg-[color:var(--theme-surface-page)]"
+        />
+      </div>
+    </section>
+  );
+}
