+++ features/shared/lib/ownerSidebarNav.ts
@@ -32,10 +32,8 @@ const OWNER_SECTION_OVERRIDES_BY_HREF: Record<string, string> = {
   "/parts": "Parts",
   "/parts/requests": "Parts",
   "/parts/receiving": "Parts",
-  "/parts/receive": "Parts",
   "/parts/inventory": "Parts",
   "/parts/po": "Parts",
-  "/parts/po/receive": "Parts",
   "/parts/movements": "Parts",
   "/parts/allocations": "Parts",
   "/parts/vendors": "Parts",
